fix: now all entries are checked; split keys check
In case there are many files in the bib file, now all of them are
checked and not only the first one (as it was).

The try catch formula is now split into 3 cases for 3 most relevant keys
that have to provided for the papers (instead of 1 block for all 3
keys).